Switch detail page to view/page format
